UIAutotest/emailUItest3/requirements.txt
allure-pytest==2.8.12
allure-python-commons==2.8.12
apipkg==1.5
appdirs==1.4.3
atomicwrites==1.3.0
attrs==19.1.0
beautifulsoup4==4.8.0
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
cssselect==1.1.0
ddt==1.2.1
docopt==0.6.2
elasticsearch5==5.5.6
execnet==1.6.1
fake-useragent==0.1.11
Faker==2.0.1
idna==2.8
importlib-metadata==0.18
lxml==4.4.1
more-itertools==7.2.0
packaging==19.0
parse==1.12.1
pipreqs==0.4.9
pluggy==0.13.1
py==1.8.0
pyee==6.0.0
PyMySQL==0.9.3
pyparsing==2.4.1.1
pypiwin32==223
pyppeteer==0.0.25
pyquery==1.4.0
pytest==5.0.0
pytest-forked==1.0.2
pytest-html==1.21.1
pytest-metadata==1.8.0
pytest-repeat==0.8.0
pytest-rerunfailures==7.0
pytest-xdist==1.29.0
python-dateutil==2.8.0
pywin32==227
PyYAML==5.1.1
requests==2.22.0
requests-html==0.10.0
selenium==3.141.0
six==1.12.0
soupsieve==1.9.3
text-unidecode==1.2
tqdm==4.36.0
urllib3==1.25.3
w3lib==1.21.0
wcwidth==0.1.7
websockets==8.0.2
yarg==0.1.9
zipp==0.5.2
